gerador-xml/requirements.txt
psycopg2-binary
xmltodict
